pyproject.toml
[tool.poetry]
name = "python_demo_hexagonal_architecture"
version = "0.1.0"
description = ""
authors = ["Алексей Лесков <alexleskov2000@gmail.com>"]
readme = "README.md"
packages = [{include = "shop", from="src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
pydantic = "^2.6.4"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
attrs = "^23.2.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
tzdata = "^2024.1"
python-json-logger = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
